requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
Pillow==10.3.0